requirements.txt
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
cycler==0.10.0
discord==1.0.1
discord.py==1.6.0
emot==2.1
Flask==1.1.2
gunicorn==20.0.4
hatesonar==0.0.7
idna==2.10
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
multidict==5.1.0
numpy==1.19.4
opencv-python==4.5.1.48
pandas==1.1.5
Pillow==8.1.0
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
soupsieve==2.1
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==1.0.1
yarl==1.6.3
